backend/requirements.txt
fastapi~=0.110.0
python-dotenv~=1.0.1
starlette~=0.36.3
uvicorn~=0.29.0
pydantic~=2.6.4

coverage~=7.4.4
httpx~=0.27.0
pytest~=8.1.1
pytest-cov~=5.0.0
pytest-html~=4.1.1
mypy~=1.9.0
tox~=4.14.0

pymongo
motor~=3.6.0
passlib~=1.7.4  
python-jose~=3.3.0
email-validator~=2.2.0
bcrypt~=4.0.1

easyocr~=1.7.2
ultralytics~=8.3.9
asyncpg~=0.29.0
Flask-mail~=0.10.0
